move plotting-related functions to dedicated module
