+++ docs/reference.rst
@@ -59,6 +59,12 @@ Image registration
 .. automodule:: dtcwt.registration
     :members:
 
+Plotting functions
+``````````````````
+
+.. automodule:: dtcwt.plotting
+    :members:
+
 Miscellaneous and low-level support functions
 `````````````````````````````````````````````
